feat!: Restart as cloudformatious-cli
The initial work on cfn-deploy led to creating the cloudformatious crate
to provide a better programmatic API for interacting with
CloudFormation. This project can now focus on the CLI side of things.

This version of the CLI is quite a bit simpler than cfn-deploy. The
biggest difference is the UI – there is no interactive terminal UI,
however the CLI is much more complete, with proper argument handling
etc. More internally, cfn-deploy supported inspecting change sets prior
to executing them, which cloudformatious currently does not.

On the flip side, this version is pretty much usable already. There's
definitely some polish needed on the output, and some of the error
handling could be simplified, but it will do the job for basic
CloudFormation deployments.