feat: Apply some basic formatting to stack events
Stack events are now in a best-effort tabular format. This is quite
accurate when applying since the appropriate size for all columns other
than status reason can be determined from the change set.

For deleting, we can still pad the resource status (since the max size
is known statically), and we can pad the resource type to at least the
length of `AWS::CloudFormation::Stack`, since we know that will appear.
We can't tell anything about the logical resource ID, since
`delete-stack` could be invoked with a stack ARN, which would be way too
long. We could try to improve the situation by describing the stack and
its resources, but that's quite an expensive operation just to improve
formatting, and would also force us to handle errors that would occur if
the stack doesn't exist which we otherwise get for free from
`cloudformatious`.