refactor!: Drop -cli suffix from binary name
This was always the intended name of the binary, but the default binary
name is exactly the name of the crate. A new name can be supplied by
explicitly defining a `[[bin]]` entry for it in `Cargo.toml`.

BREAKING CHANGE: When installed, the binary is now named
`cloudformatious`, not `cloudformatious-cli`.