refactor: Upgrade cloudformatious for status reason parsing
Status reason parsing has been upstreamed into the `cloudformatious`
crate itself, meaning we no longer have to do it here. We do still have
to convert the status reason detail into a hint, if possible.

Since `StatusReasonDetail` is a non-exhaustive enum, we still have to
deal with the possibility that the status reason is pareseable but a
hint could not be generated. To prevent this, the overall hint
generation would have to move upstream, or the `StatusReasonDetail` enum
would need to be made exhaustive, and neither is currently desireable.