feat: support SSO authentication
This uses the [`aws_sso_flow`] library to add SSO auth as a fallback if
the usual chain fails.

Unfortunately, there doesn't seem to be a way to reliably indicate that
`aws_sso_flow` should depend on `rusoto_credential` @ `0.46.0`, which we
require for the build to succeed since our rusoto deps are all 0.46.0.
As such, `cargo update` keeps trying to bump that dependency. There may
be some way around this, but in the meantime the `cargo update` step has
been made a no-op, which might slow down CI slightly.

[`aws_sso_flow`]: https://docs.rs/aws_sso_flow/