feat: add an (unstable) `package` command
This offers similar functionality as `aws cloudformation package`,
whereby local paths in certain template properties are uploaded to S3
and replaced with a reference to the uploaded file. The new
`lambda.yaml` fixture demonstrates this for a Lambda function.

Like `aws cloudformation package`, the `package` command will print the
modified template. The plan is to not have a separate command for this
however, and to instead integrate packaging with `apply-stack` (hence
the `package` command is unstable as it might not last long).

Terminologically, we say that some resource properties are
"packageable". Packageable properties are identified by resource type
and a path into the resource properties (e.g. `AWS::Lambda::Function`
and `["Code"]`, indicating that the `Code` property of
`AWS::Lambda::Function`s is packageable). Packaging a template takes
place in two phases:

1. Locating "package targets" in the template. This involves visiting
   each packageable property in the template, checking if it should be
   packaged, and extracting a `package::Target`s if so. For now,
   packageable properties will be packaged if they contain a string
   value. This may end up depending on the particular `PackageableField`
   in future.

2. "Processing" the targets, which involves building a zip archive of
   the local path, uploading the archive to S3, and writing the S3
   location back to the target property. Targets are processed
   concurrently. S3 objects are named as the MD5 checksum of the zip
   file, and the upload is not performed if a file already exists.