refactor!: require `--stack-name` in `{apply,delete}-stack`
The idea originally was to minimise the amount of configuration required
to apply a stack by inferring a name from the template name. In reality
this is probably an unhelpful and unintuitive behaviour since stack
names don't follow template file names in any but the most trivial
deployments.

BREAKING CHANGE: The `--stack-name` option is now required for the
`apply-stack` and `delete-stack` commands. The `delete-stack` command
will no longer accept a positional argument for the template path.