refactor!: make `template-path` a (required) option
Using a positional argument for the template path can easily lead to
confusion since positional arguments cannot come after multiple-valued
options (e.g.
`cloudformatious apply-stack --parameters a=1 my-template.yaml` will
result in an error because `my-template.yaml` can't be parsed as a
parameter). This is made worse by the fact that clap's generated usage
shows options before positionals (if there was a way to force
positionals to be first then this wouldn't be an issue).

BREAKING CHANGE: The template path must now be set with a
`--template-path` option.