feat: support packaging of CloudFormation stack's `TemplateURL`s
The `TemplateURL` of `AWS::CloudFormation::Stack` resources is now a
packageable property. This required adding a `strategy` field to the
`PackageableProperty` struct to indicate how the local path should be
processed, since the S3 reference for `TemplateURL` must point to a
plain file (i.e. rather than a zip file).

Technically a `zip: bool` field would have been sufficient for now, but
the `PackageStrategy` enum is more extensible and arguably clearer.