fix: use correct URI for CloudFormation::Stack S3 refs
The `AWS::CloudFormation::Stack` `TemplateURL` property must be a URL to
a file in S3. Notably, the URL must follow the appropriate canonical form
for the target bucket (e.g. path-based or virtual-host based, etc).

The current version of the SDK doesn't offer a convenient way to resolve
the correct S3 URI from configuration, but we can intercept the
`head_object` request and extract the URI from it.